==> 1_TILES_engage_survey_prepro.ipynb <==
Preprocess and clean engage_psycap data from S3 bucket

### 'results_updated' column contains answers to the engage_psycap survey questions 1-15, need to be processed according to JV instructions as follows:

relabel “1” as “location”

relabel “2” as “activity”

create ordinal variable “engage_mgt” by computing mean of items 3-5 (range 1-7) – this is the main work engagement score

create ordinal variable “psycap_mgt” by computing mean of items 6-17 (range 1-7) – this is the main psychological capital score

create ordinal variable “support_mgt” by computing mean of items 18-21 (range 1-7) – this score indicates job support

create ordinal variable “challenge_mgt” by computing mean of items 22-25 (range 1-7) – this score indicates challenge stressors (often positive)

create ordinal variable “hindrance_mgt” by computing mean of items 26-29 (range 1-7) – this score indicates hindrance stressors (often negative)


#getting and working with data
import pandas as pd
import numpy as np
import re

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

import warnings; warnings.simplefilter('ignore')
np.set_printoptions(suppress=True)

path_PF_S3 = 'C:/Users/Schindler/Documents/Schindler_Lab/Data/Clinical projects/TILES/PF_survey_fix/S3_app_surveys_updated.csv'

#read in csv from S3 bucket; psychologial felxibility EMAs in long form
data_PF_S3 = pd.read_csv(path_PF_S3)
data_PF_S3 = pd.DataFrame(data = data_PF_S3)
print(data_PF_S3.shape)

#split off the engage_psycap survey
data_PF_S3_engage_only = data_PF_S3[data_PF_S3['survey_type'] == 'engage_psycap']
print(data_PF_S3_engage_only.shape)

#each participant should have 20 survey entries
data_PF_S3_engage_only_participant_vc = data_PF_S3_engage_only['participant_id'].value_counts()
#save this info as csv to share with JV
data_PF_S3_engage_only_participant_vc.to_csv('data_PF_S3_engage_only_participant_vc.csv')

def process_engage(data):
    engage_dic = {'engage_location': np.nan,
              'engage_activity': np.nan, 
              'engage_3': np.nan,
              'engage_4': np.nan,
              'engage_5': np.nan,
              'engage_6': np.nan,
              'engage_7': np.nan,
              'engage_8': np.nan,
              'engage_9': np.nan,
              'engage_10': np.nan,
              'engage_11': np.nan,
              'engage_12': np.nan,
              'engage_13': np.nan,
              'engage_14': np.nan,
              'engage_15': np.nan,
              'engage_16': np.nan,
              'engage_17': np.nan,
              'engage_18': np.nan,
              'engage_19': np.nan,
              'engage_20': np.nan,
              'engage_21': np.nan,
              'engage_22': np.nan,
              'engage_23': np.nan,
              'engage_24': np.nan,
              'engage_25': np.nan,
              'engage_26': np.nan,
              'engage_27': np.nan,
              'engage_28': np.nan,
              'engage_29': np.nan,
              'engage_mgt': np.nan,
              'psycap_mgt': np.nan,
              'support_mgt': np.nan,
              'challenge_mgt': np.nan,
              'hindrance_mgt': np.nan}

    #determine if they answered the survey, if they did then process and update qno2_dic and q2_dic
    try:
        type(data) == str
    
        #create dic from entry
        dic = eval(data)
        
        try: #not everyone answered all questions
            engage_dic['engage_location'] = dic['1']
        except:
            pass
        try: 
            engage_dic['engage_activity'] = dic['2']
        except:
            pass
        try: 
            engage_dic['engage_3'] = dic['3']
        except:
            pass
        try: 
            engage_dic['engage_4'] = dic['4']
        except:
            pass
        try: 
            engage_dic['engage_5'] = dic['5']
        except:
            pass
        try: 
            engage_dic['engage_6'] = dic['6']
        except:
            pass
        try: 
            engage_dic['engage_7'] = dic['7']
        except:
            pass
        try: 
            engage_dic['engage_8'] = dic['8']
        except:
            pass
        try: 
            engage_dic['engage_9'] = dic['9']
        except:
            pass
        try: 
            engage_dic['engage_10'] = dic['10']
        except:
            pass
        try: 
            engage_dic['engage_11'] = dic['11']
        except:
            pass
        try: 
            engage_dic['engage_12'] = dic['12']
        except:
            pass
        try: 
            engage_dic['engage_13'] = dic['13']
        except:
            pass
        try: 
            engage_dic['engage_14'] = dic['14']
        except:
            pass
        try: 
            engage_dic['engage_15'] = dic['15']
        except:
            pass
        try: 
            engage_dic['engage_16'] = dic['16']
        except:
            pass
        try: 
            engage_dic['engage_17'] = dic['17']
        except:
            pass
        try: 
            engage_dic['engage_18'] = dic['18']
        except:
            pass
        try: 
            engage_dic['engage_19'] = dic['19']
        except:
            pass
        try: 
            engage_dic['engage_20'] = dic['20']
        except:
            pass
        try: 
            engage_dic['engage_21'] = dic['21']
        except:
            pass
        try: 
            engage_dic['engage_22'] = dic['22']
        except:
            pass
        try: 
            engage_dic['engage_23'] = dic['23']
        except:
            pass
        try: 
            engage_dic['engage_24'] = dic['24']
        except:
            pass
        try: 
            engage_dic['engage_25'] = dic['25']
        except:
            pass
        try: 
            engage_dic['engage_26'] = dic['26']
        except:
            pass
        try: 
            engage_dic['engage_27'] = dic['27']
        except:
            pass
        try: 
            engage_dic['engage_28'] = dic['28']
        except:
            pass
        try: 
            engage_dic['engage_29'] = dic['29']
        except:
            pass
        
    except:
        pass
    
    #create dataframes
    engage_df = pd.DataFrame.from_dict(engage_dic, orient='index').T

    #see instructions above for more details
    engage_df['engage_mgt'] = engage_df.loc[:, 'engage_3':'engage_5'].mean(axis = 1)
    engage_df['psycap_mgt'] = engage_df.loc[:, 'engage_6':'engage_17'].mean(axis = 1)
    engage_df['support_mgt'] = engage_df.loc[:, 'engage_18':'engage_21'].mean(axis = 1)
    engage_df['challenge_mgt'] = engage_df.loc[:, 'engage_22':'engage_25'].mean(axis = 1)
    engage_df['hindrance_mgt'] = engage_df.loc[:, 'engage_26':'engage_29'].mean(axis = 1)

    return engage_df

#create new data frame that will contain data_PF_S3_PFonly and the processed pf data for each entry
col_names = ['engage_location', 'engage_activity', 'engage_3', 'engage_4', 'engage_5', 'engage_6', 'engage_7', 'engage_8', 'engage_9', 'engage_10', 'engage_11', 'engage_12', 'engage_13', 'engage_14', 'engage_15', 'engage_16', 'engage_17', 'engage_18', 'engage_19', 'engage_20', 'engage_21', 'engage_22', 'engage_23', 'engage_24', 'engage_25', 'engage_26', 'engage_27', 'engage_28', 'engage_29', 'engage_mgt', 'psycap_mgt', 'support_mgt', 'challenge_mgt', 'hindrance_mgt']
engage_final = pd.DataFrame(columns=col_names, index=data_PF_S3_engage_only.index)
engage_final = pd.concat([data_PF_S3_engage_only, engage_final], axis = 1)

for index, row in data_PF_S3_engage_only.iterrows():
    df = process_engage(row['results_updated'])
    engage_final.loc[index, col_names] = df.values[0]

engage_final.head()

#save this info as csv to use for subsequent analysis and to share with JV and rest of group
engage_final.to_csv('engage_final.csv')
